% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_utilities.R
\name{append2H5}
\alias{append2H5}
\title{Write matrix in chunks to existing/empty HDF5 file}
\usage{
append2H5(x, h5file, printstatus = TRUE)
}
\arguments{
\item{x}{matrix}

\item{h5file}{path to the existing/empty HDF5 file}

\item{printstatus}{whether to print status}
}
\value{
Append matrix to existing/empty HDF5 file
}
\description{
Write matrix in chunks to existing/empty HDF5 file
}
\examples{
mat <- matrix(1:12, nrow=3)
rownames(mat) <- paste0("r", 1:3); colnames(mat) <- paste0("c", 1:4)
tmp_file <- tempfile(fileext=".h5")
createEmptyH5(tmp_file)
append2H5(mat, tmp_file)
h5ls(tmp_file)
}
