\name{cmap}
\alias{CMAP}
\title{cmap signature database}
\description{
  The `cmap` database represents the log_2 fold change (log2FC) of 
  12,437 genes from 1,281 compound treatments in 4 cells 
  (3,478 signatures in total). 
  
  It is stored as HDF5 backed `SummarizedExperiment` object.
}
\details{
The resource data was downloaded from the Connectivity Map project
\url{https://portals.broadinstitute.org/cmap/}, 
which is the CEL files generated by Affymetrix chips. 
The downloaded CEL files were normalized by MAS5 method. 
Then the \code{limma} package was used to do the differential expression (DE)
analysis to get the \code{logMA} matrix containing the log2FC scores. 
The code used for DE analysis is vailable at the Longevity Tools website
\url{http://girke.bioinformatics.ucr.edu/longevityTools/mydoc/mydoc_longevityTools_eDRUG_01.html}.
The `logMA` matrix was used to build the HDF5 backed 
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object, 
which is the `cmap` database.
}
\references{
  \url{http://girke.bioinformatics.ucr.edu/longevityTools/mydoc/mydoc_longevityTools_eDRUG_01.html} 
}

\seealso{
  \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
}

\keyword{datasets}

