% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineNormRes.R
\name{combineNormRes}
\alias{combineNormRes}
\title{combine normalization results from different chip types}
\usage{
combineNormRes(chiptype_dir, norm_method)
}
\arguments{
\item{chiptype_dir}{chiptype directories}

\item{norm_method}{normalization method, one of "MAS5" or "rma".}
}
\value{
data frame storing normalization values combined from chiptypes
}
\description{
Combine normalization results into a single data frame from difference
chip types
}
\examples{
# chiptype_dir <- unique(readRDS("./data/chiptype.rds"))
combineNormRes(chiptype_dir, norm_method="not run")
}
