% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_utilities.R
\name{createEmptyH5}
\alias{createEmptyH5}
\title{Create empty HDF5 file}
\usage{
createEmptyH5(h5file, delete_existing = FALSE, level = 6)
}
\arguments{
\item{h5file}{path to the HDF5 file to be created}

\item{delete_existing}{whether to delete the HDF5 if already existed}

\item{level}{The compression level used. An integer value between 0 
(no compression) and 9 (highest and slowest compression).}
}
\value{
empty HDF5 file
}
\description{
The created HDF5 file has three datasets named "assay", "colnames", 
"rownames" under root group.
}
\examples{
tmp_file <- tempfile(fileext=".h5")
createEmptyH5(tmp_file, level=6)
}
