% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gctx2h5.R
\name{gctx2h5}
\alias{gctx2h5}
\title{gctx to hdf5 file}
\usage{
gctx2h5(gctx, cid, new_cid = cid, h5file, chunksize = 5000,
  overwrite = TRUE)
}
\arguments{
\item{gctx}{character(1), path to the LINCS level5 gctx file}

\item{cid}{either a character vector or integer column indices indicating
the columns of the matrix to be parsed.}

\item{new_cid}{a character vector of the same length of cid, assigning new
column names for the matrix}

\item{h5file}{character(1), path to the destination hdf5 file}

\item{chunksize}{size of the matrix (number of columns) to be read in by chunks}

\item{overwrite}{TRUE or FALSE, whether to overwrite or append
matrix to the exsisting 'h5file'}
}
\value{
HDF5 file, representing 'lincs' database
}
\description{
Read large matrix in gctx file by chunks and write the matrix to an HDF5 file
}
\examples{
gctx <- system.file("extdata", "test_sample_n2x12328.gctx", package="signatureSearchData")
h5file <- tempfile(fileext=".h5")
gctx2h5(gctx, cid=1:2, new_cid=c('sirolimus__MCF7__trt_cp', 'vorinostat__SKB__trt_cp'), 
        h5file=h5file, chunksize=5000, overwrite=TRUE)
}
