\name{goAnno_drug}
\alias{goAnno_drug}
\title{Annotation of GO ID to drug name mappings} 
\description{
It is an intermediate file storing the GO ID to drug name mapping information
for \code{dsea_*} functions in \code{signatureSearch} package.
}

\details{
The GO ID-GENE SYMBOL mappings were converted to GO ID-drug mappings via
drug-target links in DrugBank, CLUE and STITCH databases
}
\seealso{
  \itemize{
    \item \code{\link{signatureSearch}}
    \item \code{\link[signatureSearch]{dsea_hyperG}}
    \item \code{\link[signatureSearch]{dsea_GSEA}}
  }
}

\keyword{datasets}

