% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fct.R
\name{load_sigdb}
\alias{load_sigdb}
\title{Load signature database}
\usage{
load_sigdb(ah_id_h5, ah_id_rds, db_name)
}
\arguments{
\item{ah_id_h5}{AnnotationHub id representing 'assays.h5' file from 
\code{\link[HDF5Array]{saveHDF5SummarizedExperiment}} function.}

\item{ah_id_rds}{AnnotationHub id representing 'se.rds' file from 
\code{\link[HDF5Array]{saveHDF5SummarizedExperiment}} function.}

\item{db_name}{character(1), name of the signatur database, e.g., "cmap"}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object
with `assays` slot as \code{\link[DelayedArray]{DelayedMatrix}}
}
\description{
Load signature databases (cmap etc.), which are HDF5 backed 
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object and stored
in \code{\link{AnnotationHub}}
}
\examples{
se <- load_sigdb("AH69075","AH69076","cmap")
}
\seealso{
\code{\link{cmap}}, \code{\link{lincs}}
}
