% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanExpr.R
\name{meanExpr}
\alias{meanExpr}
\title{Take mean expression values of drug treatment samples}
\usage{
meanExpr(expr_df, cmap_inst)
}
\arguments{
\item{expr_df}{data.frame with expression values in CEL samples}

\item{cmap_inst}{data.frame defining drug treatment of CEL samples}
}
\value{
data.frame with mean expression values of drug treatment in a cell
}
\description{
Take mean expression values of multiple drug treatment samples at different 
concentration and duration in the same cell as expression value of that drug
treatment in the cell.
}
\examples{
path <- system.file("extdata", "cmap_instances_02.txt", package="signatureSearchData")
cmap_inst <- read.delim(path, check.names=FALSE) 
expr_df <- as.data.frame(matrix(runif(30), ncol=5))
colnames(expr_df) <- paste0(cmap_inst$perturbation_scan_id[1:5],".CEL")
mexpr <- meanExpr(expr_df, cmap_inst[1:5,])
}
