% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCel.R
\name{normalizeCel}
\alias{normalizeCel}
\title{normalize CEL files}
\usage{
normalizeCel(chiptype_list, partition = "girkelab", rerun = TRUE)
}
\arguments{
\item{chiptype_list}{list, storing CEL files in each chiptype}

\item{partition}{partition of a computer cluster}

\item{rerun}{TRUE or FALSE, whether to run the function}
}
\value{
files storing normalized matrix in each chip type
}
\description{
Process each chip type separately running MAS5 with BiocParallel on cluster
}
\examples{
# chiptype_list <- split(names(chiptype), as.character(chiptype))
normalizeCel(chiptype_list, rerun=FALSE)
}
