% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCel.R
\name{normalizeCel_rma}
\alias{normalizeCel_rma}
\title{normalize CEL files with rma method}
\usage{
normalizeCel_rma(chiptype_list, rerun = FALSE)
}
\arguments{
\item{chiptype_list}{list, storing CEL files in each chiptype}

\item{rerun}{TRUE or FALSE, whether to run the function}
}
\value{
files storing rma normalized matrix in each chip type
}
\description{
Process each chip type separately running rma normalization
}
\details{
The rma method normalize all the CEL files in the same chip type as one batch,
it need to allocate large number of memory, 512Gb works
}
\examples{
# chiptype_list <- split(names(chiptype), as.character(chiptype))
normalizeCel_rma(chiptype_list, rerun=FALSE)
}
