% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_utilities.R
\name{readHDF5chunk}
\alias{readHDF5chunk}
\title{Read HDF5 file as SummarizedExperiment object}
\usage{
readHDF5chunk(h5file, colindex = 1:10)
}
\arguments{
\item{h5file}{character(1), path to the HDF5 file}

\item{colindex}{index of columns of the matrix to be read in}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object
}
\description{
Read in a subset of matrix from HDF5 file as 
\code{\link[SummarizedExperiment]{SummarizedExperiment}} object
}
\examples{
gctx <- system.file("extdata", "test_sample_n2x12328.gctx", package="signatureSearchData")
h5file <- tempfile(fileext=".h5")
gctx2h5(gctx, cid=1:2, new_cid=c('sirolimus__MCF7__trt_cp', 'vorinostat__SKB__trt_cp'), 
        h5file=h5file, chunksize=5000, overwrite=TRUE)
se <- readHDF5chunk(h5file, colindex=1:2)
}
