% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultisignatureSearchdata.R
\docType{data}
\name{SampleDEGdata}
\alias{SampleDEGdata}
\title{SampleDEGdata}
\format{
A data table with 19621 rows and 73 columns.
}
\usage{
SampleDEGdata
}
\description{
A data table that contains sample differential gene expression data from 14 different data sets generated from mice administered either metformin or acarbose.
}
\examples{
data(SampleDEGdata)
}
\keyword{datasets}
