% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_filter.R
\name{sig_filter}
\alias{sig_filter}
\title{LINCS signature meta data filter}
\usage{
sig_filter(meta, dose, time = "24 h")
}
\arguments{
\item{meta}{tibble, read in from LINCS signature info meta data}

\item{dose}{dose/concentration of compound treatment, need to match elements
in 'pert_idose' column of 'meta'}

\item{time}{compound treatment time, need to match elements in 'pert_itime' 
column of 'meta'}
}
\value{
tibble
}
\description{
Filter LINCS level5 signatures at specific concentation and treatment time
}
