% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineResults.R
\name{combineResults}
\alias{combineResults}
\title{combine normalization results}
\usage{
combineResults(chiptype_dir, rerun = TRUE)
}
\arguments{
\item{chiptype_dir}{chiptype directories}

\item{rerun}{TRUE or FALSE, whether to run the function}
}
\value{
files storing normalization values for each chiptype
}
\description{
Combine normalization results from same chip type in single data frame
}
\examples{
# chiptype_dir <- unique(readRDS("./data/chiptype.rds"))
combineResults(chiptype_dir, rerun=FALSE)
}
