% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCmapCEL.R
\name{getCmapCEL}
\alias{getCmapCEL}
\title{get CEL files}
\usage{
getCmapCEL(rerun = TRUE)
}
\arguments{
\item{rerun}{TRUE or FALSE, whether to download the data}
}
\value{
files
}
\description{
Download CEL files from CMap project \url{http://www.broadinstitute.org/cmap}
}
\examples{
getCmapCEL(rerun=FALSE) # set 'rerun' to TRUE if download
}
