\name{goAnno}
\alias{goAnno}
\title{Annotation of GO ID to GENE SYMBOL mappings} 
\description{
 It is an intermediate file storing the GO ID to GENE SYMBOLE mapping 
 information for \code{dtnetplot} function in \code{signatureSearch} package
}

\seealso{
  \itemize{
    \item \code{\link{signatureSearch}}
    \item \code{\link[signatureSearch]{dtnetplot}}
  }
}

\keyword{datasets}

