% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_filter.R
\name{inst_filter}
\alias{inst_filter}
\title{LINCS instance meta data filter}
\usage{
inst_filter(meta, dose = 10, dose_unit = "um", time = 24,
  time_unit = "h")
}
\arguments{
\item{meta}{tibble, read in from LINCS instance info meta data}

\item{dose}{dose/concentration of compound treatment, need to match elements
in 'pert_dose' column of 'meta'}

\item{dose_unit}{unit of dose of compound treatment, need to match elements
in 'pert_dose_unit' column of 'meta'}

\item{time}{compound treatment time, need to match elements in 'pert_time' 
column of 'meta'}

\item{time_unit}{unit of time, need to match elements in 'pert_time_unit' 
column of 'meta'}
}
\value{
tibble
}
\description{
Filter LINCS level3 instances at specific concentration and treatment time
}
\examples{
meta <- data.frame(pert_dose=c(2,4,10), pert_dose_unit="um", 
                   pert_time=24, pert_time_unit="h",
                   pert_type="trt_cp", pert_iname=c("p1","p2","p3"),
                   cell_id="MCF7")
inst_filter(meta)
}
