% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanExpr2h5.R
\name{meanExpr2h5}
\alias{meanExpr2h5}
\title{Calculate mean expression value of compound treatment samples and save to HDF5 file}
\usage{
meanExpr2h5(gctx, inst, h5file, chunksize = 2000, overwrite = TRUE)
}
\arguments{
\item{gctx}{character(1), path to the LINCS level3 gctx file}

\item{inst}{tibble, LINCS level3 instances after filtering at specific concentration and time}

\item{h5file}{character(1), path to the destination hdf5 file}

\item{chunksize}{size of the matrix (number of columns) to be saved in HDF5 file by chunks}

\item{overwrite}{TRUE or FALSE, whether to overwrite or append the
matrix to the exsisting 'h5file'}
}
\value{
HDF5 file, representing the 'lincs_expr' database
}
\description{
Calculate mean expression values of replicated samples that are treated by the same
compound in the same cell type with selected concentation and treatment time after 
filtering the LINCS level3 instances/samples. 
Then write the mean expression matrix to an HDF5 file, which is the `lincs_expr`
database.
}
\examples{
\dontrun{
meanExpr2h5(gctx="./data/GSE92742_Broad_LINCS_Level3_INF_mlr12k_n1319138x12328.gctx",
inst=inst42_filter, h5file="./data/lincs_expr.h5")
}
}
