% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLimma.R
\name{runLimma}
\alias{runLimma}
\title{DEG analysis with Limma}
\usage{
runLimma(df, comp_list, fdr = 0.05, foldchange = 1, verbose = TRUE)
}
\arguments{
\item{df}{data.frame storing normalized intensity values of samples}

\item{comp_list}{CEL file list for treatment vs. control comparisons}

\item{fdr}{cutoff of false discovery rate for defining DEGs}

\item{foldchange}{cutoff of log2 fold change for defining DEGs}

\item{verbose}{TRUE or FALSE}
}
\value{
list containing DEGs and log2FC matrix
}
\description{
DEG analysis with Limma as specified fdr and foldchange
}
\examples{
path <- system.file("extdata", "cmap_instances_02.txt", package="signatureSearchData")
cmap_inst <- read.delim(path, check.names=FALSE) 
comp_list <- sampleList(cmap_inst, myby="CMP_CELL")
df <- as.data.frame(matrix(runif(70), ncol=7))
colnames(df) <- unlist(comp_list[1])
degList <- runLimma(df, comp_list[1])  
}
