% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLimma.R
\name{runLimma}
\alias{runLimma}
\title{DEG analysis with Limma}
\usage{
runLimma(df, comp_list, fdr = 0.05, foldchange = 1, verbose = TRUE)
}
\arguments{
\item{df}{data.frame storing normalized intensity values of samples}

\item{comp_list}{CEL file list for treatment vs. control comparisons}

\item{fdr}{cutoff of false discovery rate for defining DEGs}

\item{foldchange}{cutoff of log2 fold change for defining DEGs}

\item{verbose}{TRUE or FALSE}
}
\value{
list containing DEGs and log2FC matrix
}
\description{
DEG analysis with Limma as specified fdr and foldchange
}
\examples{
a=1
# degMA <- runLimma(df, comp_list, fdr=0.10, foldchange=1, verbose=TRUE)
}
